+++ CHANGELOG.md
@@ -5,6 +5,7 @@ adheres to the format set out by [Keep a Changelog](https://keepachangelog.com/e
 
 # Unreleased
 
+- [`Feature`] Add the ability for ABI simple type conversion of entry function arguments with both remote and local ABI
 - [`Fix`] Ensure proper cleanup of response body on read error to prevent potential memory leak.
 
 # v1.5.0 (2/10/2024)