+++ CHANGELOG.md
@@ -9,6 +9,9 @@ adheres to the format set out by [Keep a Changelog](https://keepachangelog.com/e
 - Add ability to simulate for any transaction including multi-agent and fee payer
 - [`Fix`] Ensure proper cleanup of response body on read error to prevent potential memory leak.
 - [`Fix`] Fixes possible conflicts between signatures of multiple goroutines
+- [`Breaking`] Change ExpirationSeconds from int64 to uint64
+- [`Fix`] Fix a self calling function in the client
+- [`Fix`/`Breaking`] Fix fungible asset API calls that didn't return a value
 
 # v1.5.0 (2/10/2024)