+++ CHANGELOG.md
@@ -13,6 +13,7 @@ adheres to the format set out by [Keep a Changelog](https://keepachangelog.com/e
 - [`Fix`] Fix a self calling function in the client
 - [`Fix`/`Breaking`] Fix fungible asset API calls that didn't return a value
 - [`Fix`] Fix errors from failed transactions in concurrent submission
+- [`Fix`] Fix AptosRpcClient, AptosFaucetClient, AptosIndexerClient interfaces to be enforced on concrete types properly
 
 # v1.5.0 (2/10/2024)